refactor: move cart functions to separate files
